ecommerce/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
python-multipart
passlib[bcrypt]